third_party/blink/web_tests/http/tests/loading/state-object-security-exception-expected.txt

main frame - DidStartNavigation
main frame - ReadyToCommitNavigation
main frame - didCommitLoadForFrame
main frame - didFinishDocumentLoadForFrame
main frame - didHandleOnloadEventsForFrame
main frame - didFinishLoadForFrame
This test makes sure that calls to pushState() and replaceState() with URLs that violate the security origin check fail as expected.



Trying to pushState() with url http://localhost/test.html failed with exception SecurityError: Failed to execute 'pushState' on 'History': A history state object with URL 'http://localhost/test.html' cannot be created in a document with origin 'http://127.0.0.1:8000' and URL 'http://127.0.0.1:8000/loading/state-object-security-exception.html'.
History length is 1
Trying to replaceState() with url http://localhost/test.html failed with exception SecurityError: Failed to execute 'replaceState' on 'History': A history state object with URL 'http://localhost/test.html' cannot be created in a document with origin 'http://127.0.0.1:8000' and URL 'http://127.0.0.1:8000/loading/state-object-security-exception.html'.
History length is 1
Trying to pushState() with url http://localhost:8001/test.html failed with exception SecurityError: Failed to execute 'pushState' on 'History': A history state object with URL 'http://localhost:8001/test.html' cannot be created in a document with origin 'http://127.0.0.1:8000' and URL 'http://127.0.0.1:8000/loading/state-object-security-exception.html'.
History length is 1
Trying to replaceState() with url http://localhost:8001/test.html failed with exception SecurityError: Failed to execute 'replaceState' on 'History': A history state object with URL 'http://localhost:8001/test.html' cannot be created in a document with origin 'http://127.0.0.1:8000' and URL 'http://127.0.0.1:8000/loading/state-object-security-exception.html'.
History length is 1
Trying to pushState() with url http://www.webkit.org/test.html failed with exception SecurityError: Failed to execute 'pushState' on 'History': A history state object with URL 'http://www.webkit.org/test.html' cannot be created in a document with origin 'http://127.0.0.1:8000' and URL 'http://127.0.0.1:8000/loading/state-object-security-exception.html'.
History length is 1
Trying to replaceState() with url http://www.webkit.org/test.html failed with exception SecurityError: Failed to execute 'replaceState' on 'History': A history state object with URL 'http://www.webkit.org/test.html' cannot be created in a document with origin 'http://127.0.0.1:8000' and URL 'http://127.0.0.1:8000/loading/state-object-security-exception.html'.
History length is 1
Trying to pushState() with url http://www.webkit.org/ failed with exception SecurityError: Failed to execute 'pushState' on 'History': A history state object with URL 'http://www.webkit.org/' cannot be created in a document with origin 'http://127.0.0.1:8000' and URL 'http://127.0.0.1:8000/loading/state-object-security-exception.html'.
History length is 1
Trying to replaceState() with url http://www.webkit.org/ failed with exception SecurityError: Failed to execute 'replaceState' on 'History': A history state object with URL 'http://www.webkit.org/' cannot be created in a document with origin 'http://127.0.0.1:8000' and URL 'http://127.0.0.1:8000/loading/state-object-security-exception.html'.
History length is 1
Trying to pushState() with url ftp://www.webkit.org/ failed with exception SecurityError: Failed to execute 'pushState' on 'History': A history state object with URL 'ftp://www.webkit.org/' cannot be created in a document with origin 'http://127.0.0.1:8000' and URL 'http://127.0.0.1:8000/loading/state-object-security-exception.html'.
History length is 1
Trying to replaceState() with url ftp://www.webkit.org/ failed with exception SecurityError: Failed to execute 'replaceState' on 'History': A history state object with URL 'ftp://www.webkit.org/' cannot be created in a document with origin 'http://127.0.0.1:8000' and URL 'http://127.0.0.1:8000/loading/state-object-security-exception.html'.
History length is 1
Trying to pushState() with url file://anyfile.html/ failed with exception SecurityError: Failed to execute 'pushState' on 'History': A history state object with URL 'file://anyfile.html/' cannot be created in a document with origin 'http://127.0.0.1:8000' and URL 'http://127.0.0.1:8000/loading/state-object-security-exception.html'.
History length is 1
Trying to replaceState() with url file://anyfile.html/ failed with exception SecurityError: Failed to execute 'replaceState' on 'History': A history state object with URL 'file://anyfile.html/' cannot be created in a document with origin 'http://127.0.0.1:8000' and URL 'http://127.0.0.1:8000/loading/state-object-security-exception.html'.
History length is 1
